+++ example/stores/sessionStore.ts
@@ -46,7 +46,7 @@ export type SessionState =
 
 // constructor
 
-export const sessionStore = MS.ctor<'session', SessionState>(() => ({ name: 'session', initState: { status: 'init' } }))
+export const sessionStore = MS.ctor<'session', SessionState>({ name: 'session', initState: { status: 'init' } })
 
 // mutationR