+++ src/react/useMutation.ts
@@ -1,4 +1,3 @@
-import type { Lazy } from 'fp-ts/function'
 import { useCallback, useEffect, useRef } from 'react'
 import type { Subscription } from 'rxjs'
 import * as MS from '../state'
@@ -14,7 +13,7 @@ export function useMutation<
     D extends Record<K, unknown>,
     K extends never
 >(
-    s: Lazy<MS.Store<N, S>>,
+    s: MS.Store<N, S>,
     mutationL: (deps: D) => MS.Mutation<NM, P, S, SS>,
     options: MS.BaseOptions & MS.DepsOptions<D>
 ): (...payload: P) => Subscription
@@ -26,7 +25,7 @@ export function useMutation<
     SS extends S,
     D extends Record<never, unknown>
 >(
-    s: Lazy<MS.Store<N, S>>,
+    s: MS.Store<N, S>,
     mutationL: () => MS.Mutation<NM, P, S, SS>,
     options?: MS.BaseOptions
 ): (...payload: P) => Subscription
@@ -39,7 +38,7 @@ export function useMutation<
     D extends Record<K, unknown>,
     K extends never
 >(
-    s: Lazy<MS.Store<N, S>>,
+    s: MS.Store<N, S>,
     mutationL: (deps?: D) => MS.Mutation<NM, P, S, SS>,
     options?: MS.BaseOptions & Partial<MS.DepsOptions<D>>
 ): (...payload: P) => Subscription {