+++ src/react/useStore.ts
@@ -1,9 +1,8 @@
-import type { Lazy } from 'fp-ts/function'
 import { useRef } from 'react'
 import type * as MS from '../state'
 import type { storeName } from '../state/stores'
 
-export function useStore<N extends storeName, S>(s: Lazy<MS.Store<N, S>>): Lazy<MS.Store<N, S>> {
+export function useStore<N extends storeName, S>(s: MS.Store<N, S>): MS.Store<N, S> {
     const store = useRef(s)
 
     return store.current